# apps/api/openapi/caudal.yaml
openapi: 3.0.3
info:
  title: Caudal API
  version: 0.1.0
paths:
...
  /api/v1/incomes:
    get:
      operationId: list_incomes_api_v1_incomes_get
      parameters:
        - $ref: "#/components/parameters/Month"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          $ref: "#/components/responses/IncomePage"
    post:
      operationId: create_income_api_v1_incomes_post
      requestBody:
        $ref: "#/components/requestBodies/CreateIncome"
      responses:
        "201":
          $ref: "#/components/responses/Income"
  /api/v1/incomes/{income_id}:
    patch:
      operationId: update_income_api_v1_incomes__income_id__patch
      parameters:
        - $ref: "#/components/parameters/IncomeID"
      requestBody:
        $ref: "#/components/requestBodies/UpdateIncome"
      responses:
        "200":
          $ref: "#/components/responses/Income"
    delete:
      operationId: delete_income_api_v1_incomes__income_id__delete
      parameters:
        - $ref: "#/components/parameters/IncomeID"
      responses:
        "204":
          description: Successful Response
...
components:
  parameters:
    Month:
      name: month
      in: query
      required: true
      schema:
        type: string
        pattern: "^\\d{4}-\\d{2}$"
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 50
        minimum: 1
        maximum: 200
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        default: 0
        minimum: 0
...
    IncomeID:
      name: income_id
      in: path
      required: true
      schema:
        type: integer
  requestBodies:
...
    CreateIncome:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/CreateIncomeRequest"
    UpdateIncome:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/UpdateIncomeRequest"
  responses:
...
    Income:
      description: Successful Response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/IncomeResponse"
    IncomePage:
      description: Successful Response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/IncomePageResponse"
...
  schemas:
...
    CreateIncomeRequest:
      type: object
      required: [amount_cents, currency, source_id, occurred_on]
      properties:
        amount_cents: { type: integer, minimum: 1 }
        currency: { type: string, minLength: 3, maxLength: 3 }
        source_id: { type: integer, minimum: 1 }
        occurred_on: { type: string, format: date }
        note: { type: string, nullable: true }
    UpdateIncomeRequest:
      type: object
      properties:
        amount_cents: { type: integer, minimum: 1, nullable: true }
        currency: { type: string, minLength: 3, maxLength: 3, nullable: true }
        source_id: { type: integer, minimum: 1, nullable: true }
        occurred_on: { type: string, format: date, nullable: true }
        note: { type: string, nullable: true }
    IncomeResponse:
      allOf:
        - $ref: "#/components/schemas/CreateIncomeRequest"
        - type: object
          required: [id, note]
          properties:
            id: { type: integer }
    IncomePageResponse:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/IncomeResponse"
        total: { type: integer }
...